Refactor weather classes and functions
