+++ data_sources/location.py
@@ -1,0 +1,4 @@
+class Location:
+    def __init__(self, lat: float, lon: float):
+        self.lat = lat
+        self.lon = lon
